smart_planner-fixed/requirements.txt
streamlit
google-generativeai>=0.3.0
